Created new class LLVMFunctionWrapper which sole purpose is to store a pointer to an llvm::Function in order to be retrieved later on.
This class is meant to be used only by some classes that represent a key point
in the object structure to execute a given function by storing information
inside the object structure rather the current Visitor. This is done so the
information generated by a visitor can be stored in the object structure and
then retrieved by a different visitor which performs a different task based
on the information created by another visitor.