Added support and fixed bug to execute a global setup for each existing group.
The bug was that the code for each global-setup was generated the wrong way
due to the fact that the same function name was used over and over again
and we were modifying the same function over and over again.

A second change was that we execute the group setup function before visitting
the children.