Fixed bug that caused teardown functions to be generated the wrong way.
This bug was caused because the same group name was being used for the teardown
llvm functions. This was fixed by adding a new attribute in the LLVMFunctionWrapper
class that will hold the group name. Then we can use the getGroupName() to
ask the object the group it belongs to.

Due to this fixing I removed a visitor method which was being used as a workaround
for the group name.