This is the second commit of the major refactor which fixed the way we execute setup and teardown statements.
The 'before' statement maps to VisitTestSetup. The 'after' statement maps to
VisitTestTeardown. All the things done and modified in the previous two
statements are reverted in VisitTestDefinition.

BUG: I also caught an important bug in TestRunnerVisitor.h:40. We were calling
clearAllGlobalMappings() which was deleting all the global variable values!
Also freeMachineCodeForFunction() was causing some problems. I need to
investigate and understand how they work.