Fixed several memory leaks.
NOTE: There are still memory leaks. Those happen whenever a calling function
throws an exception and we have several pointers stored in a vector, thus the
vector gets destroyed but not the objected pointed to. An example of this would
be the methods:

	TestDriver::ParseMockupFixture
	TestDriver::ParseTestFixture.

The solution is to use smart pointers, that way when those pointers go out of
scope regardless of normal execution flow or an exception they will get
destroyed.