Fixed bug: Invalid memory allocation done in a different function.
Before the major refactoring took place all of the setup and teardown was being
done in one single LLVM function. For the case in which we were allocating
memory for pointer this worked just fine because the memory was being allocated
on the stack, we passed a pointer to the function and after returning the
function the memory was deallocated automatically.

For the major refactor what we did was to actually generate one llvm::Function
for each of the setup, function under test, teardown and cleanup. However the
allocation of memory was being done in the setup function properly and we
stored the memory address but after returning from that function the memory
was being deallocation, but we were still using that memory address in the
test function, causing to a segmentation fault.

The solution consisted in savingthe instructions for allocating memory in
a different vector and just insert them in the test function at a later stage
from when those instructions were generated.