Fixed bug when using BufferAlloc syntax to allocate memory for a structure.
Actually the previous commit added the instructions generated to the vector
mPtrInstructions in the setup and teardown methods, but that was not the issue.
The issue was that the method that allocated the structure called another
recursive method that was storing the generated instructions into a different
instruction vector.

Also updated the tests.