Fixed bug which would make jcut exit when an integer or float constant was passed to a pointer type.
The jcut behavior is to prevent initialization of pointer types with arbitrary
addresses. The only arbitrary address allowed to be used is 0 (NULL or nullptr)
only for the cases in which the uses wants to test error handling flows.

The other valid way to initialize pointer types is using the jcut pointer
initialization syntax [n:x].