Implemented the checking of the ExpectedExpressions for the final test result
Now we can detect when an ExpectedExpression failed, but we still need to
report which ExpectedExpression failed!

This also changed caused the ExpectedExpressions to be evaluated first, this is
because we are generating a function for the ExpextedExpression and the
code for the variable assignment and function called are created in another
function in the test-fixture one tree node above it. I may need to generate
separate functions for each function call and variable assignment, this way
the expressions will be evaluated in the same order they defined by the user.