Added support to read a CSV file and generate as many functions as there are rows in the CSV file.
The rows represent a single function under test, and each column represents
a DataPlaceholder. The restriction of this CSV file is that it has to have
M rows with N columns each.

If a given row has different column count than other rows it will be marked
as invalid CSV file.