Fixed bug which would cause many wrappers to be created.
We only need to create one wrapper for each function under test, and then
update the function pointer accordingly with the mockup functions created.
This means that for every mockup function we create 2 functions: one for
storing the address of the mockup function in the function pointer and the
second function stores the address of the actual function in the function
pointer.

For all the mockup functions created for a specific function only one
wrapper function will be created. This wrapper function just calls the
the function pointer with the passed arguments.