Fixed bug: Since we are parsing the arguments with the CommonOptionsParser constructor there is a static variable list which holds all the source files and we repeat it over and over again, the same files are duplicated. Thus we are removing the duplicated.
However, note the way the duplicated are removed, we are checking just for
the file name and not the absolute path. This is because the file which contains
only the file name it gets expanded to its absolute path after executing once,
the rest of the times gets added again without the absolute path.